requirements.txt
Django==5.2.7
djangorestframework==3.16.1
django-cors-headers==4.6.0
django-filter==24.3
Pillow==11.0.0
PyPDF2==3.0.1
drf-spectacular==0.28.0
gunicorn==21.2.0
whitenoise==6.6.0
dj-database-url==2.1.0
psycopg2-binary==2.9.9
django-storages==1.14.2
boto3==1.34.34